Add views for creating, updating and deleting categories and subcategories, and deleting shouts
- add category_create_view
- add category_update_view
- add category_delete_view
- add subcategory_create_view
- add subcategory_update_view
- add subcategory_delete_view
- add shout_delete_view
- refactor some views to use can_post, can_delete and can_edit functions for permission checks

+++ punkweb_bb/views.py
@@ -4,54 +4,46 @@ from django.contrib.auth import authenticate, get_user_model, login, logout
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponseForbidden
 from django.shortcuts import get_object_or_404, redirect, render
+from django.urls import reverse
 from django.utils import timezone
 
-from punkweb_bb.guests import guest_list
 from punkweb_bb.forms import (
     BoardProfileModelForm,
+    CategoryModelForm,
     LoginForm,
     PostModelForm,
     ShoutModelForm,
     SignUpForm,
+    SubcategoryModelForm,
     ThreadModelForm,
 )
+from punkweb_bb.guests import guest_list
 from punkweb_bb.models import Category, Post, Shout, Subcategory, Thread
 from punkweb_bb.pagination import paginate_qs
 from punkweb_bb.response import htmx_redirect
+from punkweb_bb.utils import get_unique_slug
 
 User = get_user_model()
 
 
-def index_view(request):
-    categories = Category.objects.all()
+def signup_view(request):
+    if request.user.is_authenticated:
+        return redirect("punkweb_bb:index")
 
-    recent_threads = Thread.objects.all().order_by("-created_at")[:5]
+    if request.method == "POST":
+        form = SignUpForm(request.POST)
 
-    thread_count = Thread.objects.count()
-    post_count = Post.objects.count()
+        if form.is_valid():
+            form.save()
 
-    users = User.objects.select_related("profile").all()
-    newest_user = users.order_by("-profile__created_at").first()
-
-    users_online = [user for user in users if user.profile.is_online]
-    members_online = [user for user in users_online if not user.is_staff]
-    staff_online = [user for user in users_online if user.is_staff]
-    guests_online = guest_list.length()
-    total_online = len(members_online) + len(staff_online) + guests_online
+            return redirect("punkweb_bb:login")
+    else:
+        form = SignUpForm()
 
     context = {
-        "categories": categories,
-        "recent_threads": recent_threads,
-        "thread_count": thread_count,
-        "post_count": post_count,
-        "users": users,
-        "newest_user": newest_user,
-        "members_online": members_online,
-        "staff_online": staff_online,
-        "guests_online": guests_online,
-        "total_online": total_online,
+        "form": form,
     }
-    return render(request, "punkweb_bb/index.html", context=context)
+    return render(request, "punkweb_bb/signup.html", context)
 
 
 def login_view(request):
@@ -85,24 +77,36 @@ def logout_view(request):
     return redirect("punkweb_bb:login")
 
 
-def signup_view(request):
-    if request.user.is_authenticated:
-        return redirect("punkweb_bb:index")
+def index_view(request):
+    categories = Category.objects.all()
 
-    if request.method == "POST":
-        form = SignUpForm(request.POST)
+    recent_threads = Thread.objects.all().order_by("-created_at")[:5]
 
-        if form.is_valid():
-            form.save()
+    thread_count = Thread.objects.count()
+    post_count = Post.objects.count()
 
-            return redirect("punkweb_bb:login")
-    else:
-        form = SignUpForm()
+    users = User.objects.select_related("profile").all()
+    newest_user = users.order_by("-profile__created_at").first()
+
+    users_online = [user for user in users if user.profile.is_online]
+    members_online = [user for user in users_online if not user.is_staff]
+    staff_online = [user for user in users_online if user.is_staff]
+    guests_online = guest_list.length()
+    total_online = len(members_online) + len(staff_online) + guests_online
 
     context = {
-        "form": form,
+        "categories": categories,
+        "recent_threads": recent_threads,
+        "thread_count": thread_count,
+        "post_count": post_count,
+        "users": users,
+        "newest_user": newest_user,
+        "members_online": members_online,
+        "staff_online": staff_online,
+        "guests_online": guests_online,
+        "total_online": total_online,
     }
-    return render(request, "punkweb_bb/signup.html", context)
+    return render(request, "punkweb_bb/index.html", context=context)
 
 
 def profile_view(request, user_id):
@@ -148,6 +152,72 @@ def settings_view(request):
     return render(request, "punkweb_bb/settings.html", context=context)
 
 
+@login_required(login_url="/login/")
+def category_create_view(request):
+    if not request.user.has_perm("punkweb_bb.add_category"):
+        return HttpResponseForbidden("You do not have permission to create categories.")
+
+    if request.method == "POST":
+        form = CategoryModelForm(request.POST)
+
+        if form.is_valid():
+            category = form.save(commit=False)
+            category.slug = get_unique_slug(Category, category.name)
+            category.save()
+
+            return redirect(category)
+    else:
+        form = CategoryModelForm()
+
+    context = {
+        "form": form,
+    }
+    return render(request, "punkweb_bb/category_create.html", context=context)
+
+
+@login_required(login_url="/login/")
+def category_update_view(request, category_slug):
+    if not request.user.has_perm("punkweb_bb.change_category"):
+        return HttpResponseForbidden("You do not have permission to change categories.")
+
+    category = get_object_or_404(Category, slug=category_slug)
+
+    if request.method == "POST":
+        form = CategoryModelForm(request.POST, instance=category)
+
+        if form.is_valid():
+            category = form.save()
+
+            return redirect(category)
+    else:
+        form = CategoryModelForm(instance=category)
+
+    context = {
+        "category": category,
+        "form": form,
+    }
+    return render(request, "punkweb_bb/category_update.html", context=context)
+
+
+@login_required(login_url="/login/")
+def category_delete_view(request, category_slug):
+    if not request.user.has_perm("punkweb_bb.delete_category"):
+        return HttpResponseForbidden("You do not have permission to delete categories.")
+
+    category = get_object_or_404(Category, slug=category_slug)
+
+    if request.method == "DELETE":
+        category.delete()
+
+        return htmx_redirect(reverse("punkweb_bb:index"))
+
+    context = {
+        "category": category,
+    }
+
+    return render(request, "punkweb_bb/partials/category_delete.html", context=context)
+
+
 def subcategory_view(request, subcategory_slug):
     subcategory = get_object_or_404(Subcategory, slug=subcategory_slug)
 
@@ -161,11 +231,91 @@ def subcategory_view(request, subcategory_slug):
 
 
 @login_required(login_url="/login/")
+def subcategory_create_view(request, category_slug):
+    if not request.user.has_perm("punkweb_bb.add_subcategory"):
+        return HttpResponseForbidden(
+            "You do not have permission to create subcategories."
+        )
+
+    category = get_object_or_404(Category, slug=category_slug)
+
+    if request.method == "POST":
+        form = SubcategoryModelForm(request.POST)
+
+        if form.is_valid():
+            subcategory = form.save(commit=False)
+            subcategory.category = category
+            subcategory.slug = get_unique_slug(Subcategory, subcategory.name)
+            subcategory.save()
+
+            return redirect(subcategory)
+    else:
+        form = SubcategoryModelForm()
+
+    context = {
+        "category": category,
+        "form": form,
+    }
+    return render(request, "punkweb_bb/subcategory_create.html", context=context)
+
+
+@login_required(login_url="/login/")
+def subcategory_update_view(request, subcategory_slug):
+    if not request.user.has_perm("punkweb_bb.change_subcategory"):
+        return HttpResponseForbidden(
+            "You do not have permission to change subcategories."
+        )
+
+    subcategory = get_object_or_404(Subcategory, slug=subcategory_slug)
+
+    if request.method == "POST":
+        form = SubcategoryModelForm(request.POST, instance=subcategory)
+
+        if form.is_valid():
+            subcategory = form.save()
+
+            return redirect(subcategory)
+    else:
+        form = SubcategoryModelForm(instance=subcategory)
+
+    context = {
+        "subcategory": subcategory,
+        "form": form,
+    }
+    return render(request, "punkweb_bb/subcategory_update.html", context=context)
+
+
+@login_required(login_url="/login/")
+def subcategory_delete_view(request, subcategory_slug):
+    if not request.user.has_perm("punkweb_bb.delete_subcategory"):
+        return HttpResponseForbidden(
+            "You do not have permission to delete subcategories."
+        )
+
+    subcategory = get_object_or_404(Subcategory, slug=subcategory_slug)
+
+    if request.method == "DELETE":
+        subcategory.delete()
+
+        return htmx_redirect(subcategory.category.get_absolute_url())
+
+    context = {
+        "subcategory": subcategory,
+    }
+
+    return render(
+        request, "punkweb_bb/partials/subcategory_delete.html", context=context
+    )
+
+
+@login_required(login_url="/login/")
 def thread_create_view(request, subcategory_slug):
     subcategory = get_object_or_404(Subcategory, slug=subcategory_slug)
 
-    if subcategory.staff_post_only and not request.user.is_staff:
-        return redirect(subcategory)
+    if not subcategory.can_post(request.user):
+        return HttpResponseForbidden(
+            "You do not have permission to post in this subcategory."
+        )
 
     if request.method == "POST":
         form = ThreadModelForm(request.POST)
@@ -211,7 +361,12 @@ def thread_view(request, thread_id):
 
 @login_required(login_url="/login/")
 def thread_update_view(request, thread_id):
-    thread = get_object_or_404(Thread, pk=thread_id, user=request.user)
+    thread = get_object_or_404(Thread, pk=thread_id)
+
+    if not thread.can_edit(request.user):
+        return HttpResponseForbidden(
+            "You do not have permission to change this thread."
+        )
 
     if request.method == "POST":
         form = ThreadModelForm(request.POST, instance=thread)
@@ -232,7 +387,12 @@ def thread_update_view(request, thread_id):
 
 @login_required(login_url="/login/")
 def thread_delete_view(request, thread_id):
-    thread = get_object_or_404(Thread, pk=thread_id, user=request.user)
+    thread = get_object_or_404(Thread, pk=thread_id)
+
+    if not thread.can_delete(request.user):
+        return HttpResponseForbidden(
+            "You do not have permission to delete this thread."
+        )
 
     if request.method == "DELETE":
         thread.delete()
@@ -250,8 +410,10 @@ def thread_delete_view(request, thread_id):
 def post_create_view(request, thread_id):
     thread = get_object_or_404(Thread, pk=thread_id)
 
-    if thread.is_closed:
-        return HttpResponseForbidden("Cannot add posts to a closed thread.")
+    if not thread.can_post(request.user):
+        return HttpResponseForbidden(
+            "You do not have permission to post in this thread."
+        )
 
     form = PostModelForm(request.POST)
 
@@ -266,7 +428,10 @@ def post_create_view(request, thread_id):
 
 @login_required(login_url="/login/")
 def post_update_view(request, post_id):
-    post = get_object_or_404(Post, pk=post_id, user=request.user)
+    post = get_object_or_404(Post, pk=post_id)
+
+    if not post.can_edit(request.user):
+        return HttpResponseForbidden("You do not have permission to change this post.")
 
     if request.method == "POST":
         form = PostModelForm(request.POST, instance=post)
@@ -288,7 +453,10 @@ def post_update_view(request, post_id):
 
 @login_required(login_url="/login/")
 def post_delete_view(request, post_id):
-    post = get_object_or_404(Post, pk=post_id, user=request.user)
+    post = get_object_or_404(Post, pk=post_id)
+
+    if not post.can_delete(request.user):
+        return HttpResponseForbidden("You do not have permission to delete this post.")
 
     if request.method == "DELETE":
         post.delete()
@@ -338,6 +506,25 @@ def shout_create_view(request):
             return render(
                 request, "punkweb_bb/shoutbox/shout_list.html", context=context
             )
+
+
+@login_required(login_url="/login/")
+def shout_delete_view(request, shout_id):
+    shout = get_object_or_404(Shout, pk=shout_id)
+
+    if not shout.can_delete(request.user):
+        return HttpResponseForbidden("You do not have permission to delete this shout.")
+
+    if request.method == "DELETE":
+        shout.delete()
+
+        return htmx_redirect(reverse("punkweb_bb:index"))
+
+    context = {
+        "shout": shout,
+    }
+
+    return render(request, "punkweb_bb/partials/shout_delete.html", context=context)
 
 
 def bbcode_view(request):
